+++ src/Core/Discovery/Attributes/Controller.php
@@ -6,12 +6,13 @@ namespace Picowind\Core\Discovery\Attributes;
 
 use Attribute;
 
-#[Attribute(Attribute::TARGET_CLASS)]
+#[Attribute(Attribute::TARGET_CLASS | Attribute::IS_REPEATABLE)]
 final class Controller
 {
     public function __construct(
         public string $namespace = 'picowind/v1',
         public string $prefix = '',
         public array $middleware = [],
+        public array $aliases = [],
     ) {}
 }